rabbitmq/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile rabbitmq/requirements.in
#
annotated-types==0.5.0
    # via pydantic
faker==19.6.2
    # via -r rabbitmq/requirements.in
greenlet==2.0.2
    # via sqlalchemy
pika==1.3.2
    # via -r rabbitmq/requirements.in
psycopg2-binary==2.9.7
    # via -r rabbitmq/requirements.in
pydantic==2.3.0
    # via -r rabbitmq/requirements.in
pydantic-core==2.6.3
    # via pydantic
python-dateutil==2.8.2
    # via faker
python-dotenv==1.0.0
    # via -r rabbitmq/requirements.in
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.21
    # via -r rabbitmq/requirements.in
typing-extensions==4.8.0
    # via
    #   annotated-types
    #   faker
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
